% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{omicplotr.anosim}
\alias{omicplotr.anosim}
\alias{omicplotr.clr}
\alias{omicplotr.colvec}
\alias{omicplotr.filter}
\alias{omicplotr.colouredPCA}
\alias{omicplotr.getRemovedSamples}
\alias{omicplotr.getRemovedFeatures}
\alias{omicplotr.metadataFilter}
\alias{omicplotr.permanova}
\alias{omicplotr.report}
\title{Internal omicplotR functions.}
\description{
Functions necessary for the app which are not intended to be called by the
 user.
}
\author{
Daniel Giguere
}
