% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.r
\docType{data}
\name{metadata}
\alias{metadata}
\title{Vaginal microbiome OTU table metadata}
\format{A data frame with 297 rows and 35 columns, where rows are samples
 and columns are collected metadata.}
\description{
Associated metadata to \code{\link{otu_table}}.
}
\references{
Macklaim et al (2014). Microb Ecol Health Dis.
doi: http://dx.doi.org/10.3402/mehd.v26.27799
}
\seealso{
\code{\link{otu_table}}
}
