% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicplotr.colouredPCA.r
\name{omicplotr.colouredPCA}
\alias{omicplotr.colouredPCA}
\title{Plot coloured biplot.}
\usage{
omicplotr.colouredPCA(data, colourvector, scale = 0, arrows = TRUE,
  taxonomy = NULL, show.taxonomy = FALSE, tax.level = 6,
  removenames = FALSE, names.cex = 1, points.cex = 0.8,
  main = "Principal\\n     Component Analysis Biplot")
}
\arguments{
\item{data}{"prcomp" object generated by \code{omicplotr.clr}.}

\item{colourvector}{List of colours generated by \code{omciplotr.colvec}.}

\item{scale}{0 = relationship between samples, 1 = relationship between rows
default 0.}

\item{arrows}{Show PCA arrows, default TRUE.}

\item{taxonomy}{Taxonomy column from \code{omicplotr.filter} to plot points
as taxonomic identifiers.}

\item{show.taxonomy}{Show taxonomy as points, default 0. Required input is
x$taxonomy.}

\item{tax.level}{Numeric taxonomic level (kingdom =1, phylum =2...) to show
if \code{show.taxonomy = TRUE}, default 6 (genus).}

\item{removenames}{Replace sample names with points, default FALSE.}

\item{names.cex}{The size of sample names.}

\item{points.cex}{The size of points representing features (text size for
taxonomic levels if \code{show.taxonomy = TRUE}).}

\item{main}{main Title for plot.}
}
\description{
Create a coloured PCA biplot from a \code{prcomp} object
 generated by \code{omicplotr.clr}.
}
\details{
Coloured PCA biplot is plotted using \code{coloredBiplot} from
 \code{zCompositions}. \code{omicplotr.colvec} generates a list of colours
 to be used when plotting the sample names.
}
\seealso{
\code{\link{omicplotr.colvec}}
}
