% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicplotr.colvec.r
\name{omicplotr.colvec}
\alias{omicplotr.colvec}
\title{Generate colour vector from metadata}
\usage{
omicplotr.colvec(data, meta, column, type = 3)
}
\arguments{
\item{data}{prcomp objected generated by \code{omicplotr.clr}.}

\item{meta}{metadata associated with data.}

\item{column}{column by which to generate a unique colour for each unique
value.}

\item{type}{colouring type for data. 1 = numeric (continuous), 2 = quartile,
3 = categorical (numeric or nonnumeric), default 3.}
}
\description{
Generates vector of colours from metadata file to be used to colour sample
names in \code{omicplotr.colouredPCA} using \code{xlabs.col}.
}
