% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicplotr.getRemovedSamples.r
\name{omicplotr.getRemovedSamples}
\alias{omicplotr.getRemovedSamples}
\title{Display removed samples.}
\usage{
omicplotr.getRemovedSamples(x, x.clr)
}
\arguments{
\item{x}{Unfiltered data.}

\item{x.clr}{Output from \code{omicplotr.clr}.}
}
\description{
Generates table of removed samples (columns) by \code{omicplotr.filter}.
}
\seealso{
\code{\link{omicplotr.clr}}
}
\author{
Daniel Giguere
}
