% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicplotr.metadataFilter.r
\name{omicplotr.metadataFilter}
\alias{omicplotr.metadataFilter}
\title{Filter data by selected metadata.}
\usage{
omicplotr.metadataFilter(data, meta, column, values)
}
\arguments{
\item{data}{unfiltered data.}

\item{meta}{unfiltered metadata.}

\item{column}{column name from metadata.}

\item{values}{list of values to keep from metadata column.}
}
\description{
Filters data by choosing values to remain using associated metadata.
}
