% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicplotr.run.R
\name{omicplotr.run}
\alias{omicplotr.run}
\title{Run \code{omicplotR} \code{Shiny} app.}
\usage{
omicplotr.run()
}
\description{
Launches \code{Shiny} for \code{omicplotR} in default browser. This
 graphical user interface facilitates visual exploration of datasets for
 both new and experienced R users alike.
}
\value{
There is no value!
}
\examples{
# To run omicplotR, use `omicplotr.run()` function. This will open the shiny
# app in your default browser.

# omicplotr.run()

# After running omicplotr.run(), you can filter your data and make a PCA
# biplot. Internal functions are not mean to be called by the user.

data(otu_table)
d.filt <- omicplotr.filter(otu_table)

}
\author{
Daniel Giguere
}
