% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otu_table.r
\docType{data}
\name{otu_table}
\alias{otu_table}
\title{Vaginal microbiome OTU table}
\format{A data frame with 77 rows and 298 columns. The last column contains
 taxonomic information for each OTU. Samples are columns, OTUs are rows.}
\usage{
data(otu_table)
}
\description{
This dataset contains a count table of operational taxonomic units (OTUs)
 from a 16s rRNA gene sequencing experiment characterising changes in
 bacterial microbiota from oral antimicrobial or probiotic treatments.
 Associated metadata is contained in \code{\link{metadata}}.
}
\references{
Macklaim et al (2014). Microb Ecol Health Dis.
doi: http://dx.doi.org/10.3402/mehd.v26.27799
}
