% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{Internal}
\alias{omicplotr.anosim}
\alias{omicplotr.clr}
\alias{omicplotr.colvec}
\alias{omicplotr.filter}
\alias{omicplotr.colouredPCA}
\alias{omicplotr.getRemovedSamples}
\alias{omicplotr.getRemovedFeatures}
\alias{omicplotr.metadataFilter}
\alias{omicplotr.permanova}
\alias{omicplotr.report}
\title{Internal omicplotR functions.}
\description{
Internal functions for the app which are not intended to be called by the
 user. Please refer to \link{omicplotr.run}.
}
\examples{
# These functions are not meant to be called by the user. However, some
# functions can be useful.

data(otu_table)
# Filter out your OTU table to remove rows that sum to less than 10.

d <- omicplotr.filter(otu_table, min.reads = 10)

}
\author{
Daniel Giguere, Greg Gloor.
}
