% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicplotr.R
\docType{package}
\name{omicplotR-package}
\alias{omicplotR}
\alias{omicplotR-package}
\title{\code{omicplotR}}
\description{
A Shiny app for visual exploration of omic datasets as
 compositions, and differential abundance analysis using ALDEx2. Useful for
 exploring RNA-seq, meta-RNA-seq, 16s rRNA gene sequencing and more with
 visualizations such as principal component analysis biplot (coloured using
 metadata for visualizing each variable), association plots (igraph),
 dendrograms and stacked bar plots, and effect plots (ALDEx2). Input is a
 table of counts and metadata file (if metadata exists), with options to
 filter data by count or by metadata to remove low counts, or to visualize
 select samples according to selected metadata. The Shiny app's graphical
 user interface facilitates visual exploration for new and experienced
 users of R alike.
}
\details{
To launch the \code{shiny} app in your default browser, use
 \code{omicplotr.run}.
}
\seealso{
\code{\link{omciplotr.filter}}, \code{\link{omciplotr.clr}},
 \code{\link{omciplotr.clr}}, \code{\link{omciplotr.colvec}},
 \code{\link{omciplotr.colouredPCA}}, \code{\link{omciplotr.run}}.
}
\author{
Daniel Giguere
}
