% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicplotr.anosim.r
\name{omicplotr.anosim}
\alias{omicplotr.anosim}
\title{Calculate \code{anosim} between conditions.}
\usage{
omicplotr.anosim(x, conds)
}
\arguments{
\item{x}{Filtered data output from \code{omicplotr.filter}. Columns must
in ordered of condition being tested (i.e., first 4 columns condition x and
next 4 columns condition 4).}

\item{conds}{a character vector of group labels.}
}
\description{
Wrapper for \code{vegan} package to calculate anosim p-value between
 conditions from cenetered-log ratio values between two conditions.
}
