% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicplotr.colvec.r
\name{omicplotr.colvec}
\alias{omicplotr.colvec}
\title{Generate colour vector from metadata}
\usage{
omicplotr.colvec(data, meta, column, type = 3)
}
\arguments{
\item{data}{"prcomp" objected generated by \code{omicplotr.clr}.}

\item{meta}{Metadata associated with data.}

\item{column}{Column by which to generate a unique colour for each unique
value.}

\item{type}{Colouring type for data. 1 = numeric (continuous), 2 = quartile,
3 = categorical (numeric or nonnumeric), default 3.}
}
\description{
Create a vector of colours to be used to colour sample names in
\code{omicplotr.colouredPCA}.
}
\details{
The colour vector is used to colour the sample names in the biplot
 generated by \code{omicplotr.colouredPCA}. This allows exploring the
 relationship between each variable from associated metadata. A maximum
 of 10 categorical variables colours can be created, however numerical
 variables will have a gradient generated when \code{type = 1}.
}
\seealso{
\code{\link{omicplotr.colouredPCA}}
}
\author{
Daniel Giguere
}
