% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicplotr.metadataFilter.r
\name{omicplotr.metadataFilter}
\alias{omicplotr.metadataFilter}
\title{Filter data by selected metadata.}
\usage{
omicplotr.metadataFilter(data, meta, column, values)
}
\arguments{
\item{data}{Unfiltered data.}

\item{meta}{Unfiltered metadata.}

\item{column}{Column name from metadata. Must exactly match.}

\item{values}{List of values to keep from metadata column. Must exactly match
the values in the metadata (no extra characters - including spaces).}
}
\description{
Filters samples in data to remain using values from associated metadata.
}
\details{
This filter must be applied before filtering by count using
 \code{omicplotr.filter} or generating \code{prcomp} object with
 \code{omicplotr.clr}.
}
\seealso{
\code{\link{omicplotr.filter}}, \code{\link{omicplotr.clr}}.
}
\author{
Daniel Giguere
}
