% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicplotr.run.r
\name{omicplotr.run}
\alias{omicplotr.run}
\title{Run \code{omicplotR} \code{Shiny} app.}
\usage{
omicplotr.run()
}
\description{
Launches \code{Shiny} for \code{omicplotR} in default browser. This
 graphical user interface facilitates visual exploration of datasets for
 both new and experienced R users alike.
}
\author{
Daniel Giguere
}
