% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omicplotr.run.R
\name{omicplotr.run}
\alias{omicplotr.run}
\title{Run \code{omicplotR} \code{Shiny} app.}
\usage{
omicplotr.run()
}
\description{
Launches \code{Shiny} for \code{omicplotR} in default browser. This
 graphical user interface facilitates visual exploration of datasets for
 both new and experienced R users alike.
}
\examples{
# To run omicplotR, use `omicplotr.run()` function. This will open the shiny
# app in your default browser.

omicplotr.run()

}
\author{
Daniel Giguere
}
